requirements.txt
streamlit==1.28.1
folium==0.14.0
streamlit-folium==0.15.0
pandas==2.1.1
numpy==1.24.3
plotly==5.17.0
geopandas==0.14.0
shapely==2.0.2
fiona==1.9.5
Pillow==10.0.1
pyproj==3.6.1
streamlit-chat==0.1.1
streamlit-option-menu==0.3.6
